Add Nucl. support for index

+++ data/createindex.sh
@@ -15,7 +15,7 @@ notExists() {
 [ ! -d "$2" ] &&  echo "tmp directory $2 not found!" && mkdir -p "$2";
 
 INPUT="$1"
-if [ -n "$NUCL" ]; then
+if [ -n "$TRANSLATED" ]; then
     # 1. extract orf
     if notExists "$2/orfs"; then
         # shellcheck disable=SC2086
@@ -37,6 +37,47 @@ if [ -n "$NUCL" ]; then
         echo "Remove temporary files"
         rm -f "$2/orfs" "$2/orfs.index" "$2/orfs.dbtype"
         rm -f "$2/orfs_aa" "$2/orfs_aa.index" "$2/orfs_aa.dbtype"
+        rm -f "$2/createindex.sh"
+    fi
+elif [ -n "$NUCL" ]; then
+      # 1. extract orf
+    if notExists "$2/nucl_split_seq"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" splitsequence "$INPUT" "$2/nucl_split_seq" $SPLIT_SEQ_PAR \
+            || fail "splitsequence died"
+    fi
+
+    if notExists "$2/nucl_split_seq_rev"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" extractframes "$2/nucl_split_seq" "$2/nucl_split_seq_rev" $EXTRACT_FRAMES_PAR  \
+            || fail "Extractframes died"
+    fi
+
+    # shellcheck disable=SC2086
+    "$MMSEQS" $INDEXER "$2/nucl_split_seq_rev" "$INPUT" $INDEX_PAR \
+        || fail "indexdb died"
+
+    if [ -n "$REMOVE_TMP" ]; then
+        echo "Remove temporary files"
+        rm -f "$2/nucl_split_seq" "$2/nucl_split_seq.index" "$2/nucl_split_seq.dbtype"
+        rm -f "$2/nucl_split_seq_rev" "$2/nucl_split_seq_rev.index" "$2/nucl_split_seq_rev.dbtype"
+        rm -f "$2/createindex.sh"
+    fi
+elif [ -n "$LIN_NUCL" ]; then
+      # 1. extract orf
+    if notExists "$2/nucl_split_seq"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" splitsequence "$INPUT" "$2/nucl_split_seq" $SPLIT_SEQ_PAR \
+            || fail "splitsequence died"
+    fi
+
+    # shellcheck disable=SC2086
+    "$MMSEQS" $INDEXER "$2/nucl_split_seq" "$INPUT" $INDEX_PAR \
+        || fail "indexdb died"
+
+    if [ -n "$REMOVE_TMP" ]; then
+        echo "Remove temporary files"
+        rm -f "$2/nucl_split_seq" "$2/nucl_split_seq.index" "$2/nucl_split_seq.dbtype"
         rm -f "$2/createindex.sh"
     fi
 else
